pounder: warn only if fifo more than half full
In regular operation when the process task frequently queues large ~14 byte write,
and if there is a miniconf update, the second will see a non-empty fifo.
It only becomes problematic if this would lead to
significant stalls. As long as updates are smaller than half the fifo
level, this is benign.